added orders interface

+++ rest_app/views/user_view.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, render_template, redirect, url_for, flash
 from flask_login import login_required
-from rest_app.service.user_service import form_user_data_parser, update_user, add_user
+from rest_app.service.user_service import form_user_data_parser, update_user, add_user, get_all_users_by_is_admin_filter
 from rest_app.service.address_service import add_address, address_data_form_parser
 from rest_app.service.common_services import delete_row_by_id
 from rest_app.forms.personal_info_forms import UpdateProfileForm, AddressForm
@@ -36,14 +36,21 @@ def update_user_details(user_id):
         args = form_user_data_parser().parse_args()
         update_user(user_id, **args)
         flash('User information was updated', 'success')
+        return redirect(url_for('user.user_detail', user_id=user_id))
     elif address_form.submit_address.data and address_form.validate():
         args = address_data_form_parser().parse_args()
         add_address(user_id=user_id, **args)
         flash('Address information has been updated', 'success')
+        return redirect(url_for('user.user_detail', user_id=user_id))
 
     populate_personal_info(address_form, profile_form, user_id)
 
-    return redirect(url_for('user.user_detail', user_id=user_id))
+    return render_template(
+        'personal_info.html',
+        profile_form=profile_form,
+        address_form=address_form,
+        user_id=user_id
+    )
 
 
 @user.route('/create', methods=['GET', 'POST'])
@@ -65,8 +72,14 @@ def create_user():
 @login_required
 @admin_required
 def delete_user(user_id):
-    delete_row_by_id(User, user_id)
-    flash('User was successfully deleted', 'success')
+    user = User.query.filter_by(id=user_id).first()
+
+    if not user.is_admin:
+        delete_row_by_id(User, user_id)
+        flash('User was successfully deleted', 'success')
+    else:
+        if len(get_all_users_by_is_admin_filter(True).all()) > 1:
+            flash('You are the only admin. Please create a new one and after repeat', 'danger')
 
     return redirect(url_for('admin.admin_main'))
 
@@ -85,4 +98,3 @@ def populate_personal_info(address_form, profile_form, user_id):
         address_form.change_address_values(addresses.pop())
 
     profile_form.change_user_form_values(user_id)
-
